+++ CHANGELOG.md
@@ -4,15 +4,30 @@
 
 ### Added
 
+- Add support for airflow 3.1.6 ([#742]).
+- Add operator versioning ([#725]).
+  - GitSync considered for v1alpha1 and v1alpha2
 - Support objectOverrides using `.spec.objectOverrides`.
   See [objectOverrides concepts page](https://docs.stackable.tech/home/nightly/concepts/overrides/#object-overrides) for details ([#726]).
+
+### Changed
+
+- Gracefully shutdown all concurrent tasks by forwarding the SIGTERM signal ([#741]).
+- Bump testing-tools to `0.3.0-stackable0.0.0-dev` ([#733]).
 
 ### Fixed
 
 - Default `API_WORKERS` to 1 (instead of letting Airflow default to 4) to prevent crashloop and update/correct docs to reflect this  ([#727]).
+- Prevent unnecessary Pod restarts when initially creating an AirflowCluster.
+  This is achieved by applying the StatefulSet after all ConfigMaps and Secrets that it mounts ([#734]).
 
+[#725]: https://github.com/stackabletech/airflow-operator/pull/725
 [#726]: https://github.com/stackabletech/airflow-operator/pull/726
 [#727]: https://github.com/stackabletech/airflow-operator/pull/727
+[#733]: https://github.com/stackabletech/airflow-operator/pull/733
+[#734]: https://github.com/stackabletech/airflow-operator/pull/734
+[#741]: https://github.com/stackabletech/airflow-operator/pull/741
+[#742]: https://github.com/stackabletech/airflow-operator/pull/742
 
 ## [25.11.0] - 2025-11-07